requirements.txt
# PyTorch and torchvision
torch>=2.0.0
torchvision>=0.15.0

# Data processing
pandas>=1.5.0
numpy>=1.23.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
opencv-python>=4.7.0

# Model utilities
timm>=0.9.0
torchinfo>=1.8.0

# Logging and experiment tracking
tensorboard>=2.12.0
PyYAML>=6.0

# Grad-CAM
grad-cam>=1.4.0

# Scientific computing
scikit-learn>=1.2.0
scipy>=1.10.0

# Progress bars
tqdm>=4.65.0